requirements.txt
grpcio
grpcio-tools
cryptography
pytest
sqlalchemy
alembic
fastapi
uvicorn[standard]
websockets
python-multipart
passlib[bcrypt]
pyjwt
kafka-python
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-grpc
opentelemetry-exporter-otlp
opentelemetry-instrumentation-sqlalchemy
prometheus-client
apscheduler
clickhouse-driver
py_webauthn